Remove AbstractUser
