--- config/doctrine/User/AbstractUser.orm.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
-                  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                  xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
-                          https://www.doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-
-    <mapped-superclass name="App\Entity\User\AbstractUser">
-        <field name="username" type="string" length="255" unique="true"/>
-        <field name="email" type="string" length="255"/>
-        <field name="externalId" type="string" length="64" unique="true"/>
-
-        <one-to-one field="permissions" target-entity="App\Entity\Permissions\Permissions" orphan-removal="true">
-            <join-column name="permissions_id" referenced-column-name="id" nullable="false"/>
-            <cascade>
-                <cascade-persist/>
-            </cascade>
-        </one-to-one>
-
-        <field name="avatarHash" type="string" length="512" nullable="true"/>
-
-        <indexes>
-            <!-- This is needed because child index mappings override parent index mappings -->
-            <index columns="created_at"/>
-            <index columns="last_updated_at"/>
-
-            <index columns="username"/>
-            <index columns="external_id"/>
-        </indexes>
-    </mapped-superclass>
-
-</doctrine-mapping>
