+++ config/doctrine/User/User.orm.xml
@@ -5,7 +5,27 @@
                           https://www.doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
 
     <entity name="App\Entity\User\User" table="users" repository-class="App\Repository\UserRepository">
-        <one-to-many field="ownedModLists" target-entity="App\Entity\ModList\ModList" mapped-by="owner"/>
+        <field name="username" type="string" length="255" unique="true"/>
+        <field name="email" type="string" length="255"/>
+        <field name="externalId" type="string" length="64" unique="true"/>
+
+        <one-to-one field="permissions" target-entity="App\Entity\Permissions\Permissions" orphan-removal="true">
+            <join-column name="permissions_id" referenced-column-name="id" nullable="false"/>
+            <cascade>
+                <cascade-persist/>
+            </cascade>
+        </one-to-one>
+
+        <field name="avatarHash" type="string" length="512" nullable="true"/>
+
+        <indexes>
+            <!-- This is needed because child index mappings override parent index mappings -->
+            <index columns="created_at"/>
+            <index columns="last_updated_at"/>
+
+            <index columns="username"/>
+            <index columns="external_id"/>
+        </indexes>
     </entity>
 
 </doctrine-mapping>
