+++ src/Entity/User/Traits/UserInterfaceTrait.php
@@ -4,8 +4,15 @@ declare(strict_types=1);
 
 namespace App\Entity\User\Traits;
 
+use App\Security\Enum\RoleEnum;
+
 trait UserInterfaceTrait
 {
+    public function getRoles(): array
+    {
+        return [RoleEnum::ROLE_USER];
+    }
+
     public function getPassword(): ?string
     {
         return null;
